+++ src/main/java/br/com/zalf/prolog/webservice/raizen/produtividade/model/insert/RaizenProdutividadeItemInsert.java
@@ -14,8 +14,8 @@ public class RaizenProdutividadeItemInsert {
     private BigDecimal valor;
     private String usina;
     private String fazenda;
-    private double raio;
-    private double toneladas;
+    private BigDecimal raio;
+    private BigDecimal toneladas;
     private Long codEmpresa;
     private Long cpfMotorista;
     private LocalDate dataViagem;
@@ -65,19 +65,19 @@ public class RaizenProdutividadeItemInsert {
         this.fazenda = fazenda;
     }
 
-    public double getRaio() {
+    public BigDecimal getRaio() {
         return raio;
     }
 
-    public void setRaio(final double raio) {
+    public void setRaio(final BigDecimal raio) {
         this.raio = raio;
     }
 
-    public double getToneladas() {
+    public BigDecimal getToneladas() {
         return toneladas;
     }
 
-    public void setToneladas(final double toneladas) {
+    public void setToneladas(final BigDecimal toneladas) {
         this.toneladas = toneladas;
     }
